fix(maintenance): plan listesi 500 fix (int4->Integer) + geofence panel cakisma + form min-w-0
GET /maintenance/plans interval_km int4 kolonu Long.class ile okunuyordu; geofence paneli floatControls ile ust uste biniyordu (top 18.5rem + z-50); bakim form sayi inputlari min-w-0 ile dar kolonda tasmayi onler

+++ vts-api-gateway/src/main/java/com/fleet/vts/gateway/web/MaintenanceController.java
@@ -201,7 +201,7 @@ public class MaintenanceController {
                         row.put("vehicleId", rs.getLong("vehicle_id"));
                         row.put("plate", rs.getString("plate"));
                         row.put("odometerKm", rs.getLong("odometer_km"));
-                        row.put("intervalKm", rs.getObject("interval_km", Long.class));
+                        row.put("intervalKm", rs.getObject("interval_km", Integer.class));
                         row.put("intervalDays", rs.getObject("interval_days", Integer.class));
                         row.put("nextDueKm", rs.getObject("next_due_km", Long.class));
                         OffsetDateTime dueAt = rs.getObject("next_due_at", OffsetDateTime.class);
